Add files via upload
GOMC examples utilizing MoSDeF as the framework to build the GOMC forcefield file, psf, and pdb files.  The GOMC configuration, .conf,  files were manually created and included in these examples, so the user can run the GOMC simulation.  The GOMC configuration file shall be generated via MoSDeF in the future.